refactor: simplify code & remove unused methods
