+++ packages/system/src/index.js
@@ -1,6 +1,5 @@
 export * from './style'
 export * from './styles/index'
-export * from './cx'
 export * from './th'
 export * from './variant'
 export * from './breakpoints'
